+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # SBuildr Changelog
 Dates are in YYYY-MM-DD format.
 
+## vNext ()
+- Adds `depends` argument to `Project.library()`, `Project.executable()`, `Project.test()`, and `Project.interfaces()`. This can be used to specify any dependencies not captured in the `libs` argument - for example, header-only packages.
+
 ## v0.5.0 (2019-08-21)
 - `tests` in CLI now runs all profiles by default.
 - `tests` now displays a summary at the end.